### 2025-08-25 #8860 @editorialbot
[PRE REVIEW]: Shekar: A Python Toolkit for Persian Natural Language Processing
**Submitting author:** @amirivojdan (<a href="http://orcid.org/0000-0003-3741-3979">Ahmad Amirivojdan</a>)
**Repository:** https://github.com/amirivojdan/shekar
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.24
