### 2025-08-26 41c811f
Update paper.bib
