+++ paper.bib
@@ -17,7 +17,8 @@
   title={naab: A ready-to-use plug-and-play corpus for Farsi},
   author={Sabouri, Sadra and Rahmati, Elnaz and Gooran, Soroush and Sameti, Hossein},
   journal={arXiv preprint arXiv:2208.13486},
-  year={2022}
+  year={2022},
+  doi={10.22034/jaiai.2024.480062.1016}
 }
 
 @inproceedings{qasemizadeh2006adaptive,
@@ -26,7 +27,8 @@
   booktitle={2006 ieee conference on cybernetics and intelligent systems},
   pages={1--6},
   year={2006},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/ICCIS.2006.252325}
 }
 
 @inproceedings{eslami2004persian,
@@ -40,21 +42,24 @@
   title={SentencePiece: A simple and language independent subword tokenizer and detokenizer for neural text processing},
   author={Kudo, Taku and Richardson, John},
   journal={arXiv preprint arXiv:1808.06226},
-  year={2018}
+  year={2018},
+  doi={10.48550/arXiv.1808.06226}
 }
 
 @article{rasooli2020persian,
   title={The Persian dependency treebank made universal},
   author={Rasooli, Mohammad Sadegh and Safari, Pegah and Moloodi, Amirsaeid and Nourian, Alireza},
   journal={arXiv preprint arXiv:2009.10205},
-  year={2020}
+  year={2020},
+  doi={10.48550/arXiv.2009.10205}
 }
 
 @article{lan2019albert,
   title={Albert: A lite bert for self-supervised learning of language representations},
   author={Lan, Zhenzhong and Chen, Mingda and Goodman, Sebastian and Gimpel, Kevin and Sharma, Piyush and Soricut, Radu},
   journal={arXiv preprint arXiv:1909.11942},
-  year={2019}
+  year={2019},
+  doi={10.48550/arXiv.1909.11942}
 }
 
 @article{rose2010automatic,
@@ -63,5 +68,6 @@
   journal={Text mining: applications and theory},
   pages={1--20},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/9780470689646.ch1}
 }
