### 2025-08-27 04d207c
Fix morph spacing correction
