+++ shekar/data/__init__.py
@@ -39,11 +39,27 @@ numbers = persian_digits + english_digits + arabic_digits
 non_left_joiner_letters = "دۀذاأآورژز"
 
 morph_suffixes = [
-    "ام", "ات", "اش", "مان", "تان", "شان",
-    "ای", "تر", "تری", "ترین", "ها", "های", "هایی",
-    "هایم", "هایت", "هایش", "هایمان", "هایتان", "هایشان",
+    "ام",
+    "ات",
+    "اش",
+    "مان",
+    "تان",
+    "شان",
+    "ای",
+    "تر",
+    "تری",
+    "ترین",
+    "ها",
+    "های",
+    "هایی",
+    "هایم",
+    "هایت",
+    "هایش",
+    "هایمان",
+    "هایتان",
+    "هایشان",
 ]
- 
+
 suffixes = [
     "گر",
     "گری",
