### 2025-08-28 3aca9e6
Enhance ZWNJ and whitespace handling
