+++ shekar/preprocessing/filters/non_persian_letter_filter.py
@@ -42,10 +42,7 @@ class NonPersianLetterFilter(BaseTextTransform):
         super().__init__()
 
         self.characters_to_keep = (
-            data.persian_full_set
-            + data.spaces
-            + data.persian_digits
-            + data.punctuations
+            data.persian_letters + data.spaces + data.persian_digits + data.punctuations
         )
 
         if keep_diacritics:
