### 2025-08-29 9ee3c01
Add batching support for large input texts
