### 2025-08-29 fb7fbee
Add global mean pooling for large text
