### 2025-08-29 791185a
Add NER for large text
