### 2025-09-03 a228794
Change default YaNormalizer style to "Joda"
