+++ tests/unit/preprocessing/test_preprocessors.py
@@ -115,13 +115,13 @@ def test_remove_extra_spaces():
 
 
 def test_ya_normalizer():
-    ya_normalizer = YaNormalizer()
+    ya_normalizer = YaNormalizer(style="standard")
 
     input_text = "خانه‌ی ما"
     expected_output = "خانۀ ما"
     assert ya_normalizer(input_text) == expected_output
 
-    ya_normalizer = YaNormalizer(style="joda")
+    ya_normalizer = YaNormalizer()
     input_text = "خانۀ ما"
     expected_output = "خانه‌ی ما"
     assert ya_normalizer(input_text) == expected_output
