### 2025-09-03 d36c417
Add YaNormalizer to normalization pipeline
