### 2025-09-20 de8cf01
add basic shekar-cli
