+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "shekar"
-version = "0.1.26"
+version = "0.1.27"
 description = "Simplifying Persian NLP for Modern Applications"
 readme = "README.md"
 license = "MIT"
@@ -12,6 +12,8 @@ authors = [{ name = "Ahmad Amirivojdan", email = "amirivojdan@gmail.com" }]
 requires-python = ">=3.10"
 dependencies = [
     "arabic-reshaper>=3.0.0",
+    "chardet>=5.2.0",
+    "click>=8.1.8",
     "emoji>=2.14.1",
     "numpy>=1.24.4",
     "onnxruntime>=1.22.1",
@@ -19,6 +21,7 @@ dependencies = [
     "python-bidi>=0.6.6",
     "regex>=2024.11.6",
     "tokenizers>=0.21.2",
+    "tqdm>=4.67.1",
     "wordcloud>=1.9.4",
 ]
 
@@ -52,6 +55,9 @@ dev = [
     "termynal>=0.13.0",
     "ruff>=0.12.5",
 ]
+
+[project.scripts]
+shekar = "shekar.cli:main"
 
 [project.urls]
 Homepage = "https://github.com/amirivojdan/shekar"
