### 2025-09-21 29ba79e
add wordcloud command
