+++ shekar/cli.py
@@ -3,17 +3,20 @@ import importlib.metadata as md
 from pathlib import Path
 import click
 from tqdm import tqdm
+from collections import Counter
 
 
 # ---------- helpers ----------
 def _detect_encoding(path: Path) -> str:
     try:
         import chardet
+
         with path.open("rb") as f:
             raw = f.read(50000)
         return chardet.detect(raw).get("encoding") or "utf-8"
     except Exception:
         return "utf-8"
+
 
 def _iter_lines(input_path: Path | None, text: str | None, encoding: str | None):
     if text is not None:
@@ -27,8 +30,9 @@ def _iter_lines(input_path: Path | None, text: str | None, encoding: str | None)
         for line in f:
             yield line.rstrip("\r\n")
 
+
 def _open_out(output: Path | None):
-    return (output.open("w", encoding="utf-8", newline="\n") if output else None)
+    return output.open("w", encoding="utf-8", newline="\n") if output else None
 
 
 # ---------- root ----------
@@ -36,10 +40,12 @@ def _open_out(output: Path | None):
 def cli():
     pass
 
+
 @cli.command()
 def version():
     """Show installed Shekar version."""
     click.echo(md.version("shekar"))
+
 
 @cli.command()
 def info():
@@ -49,34 +55,181 @@ def info():
 
 # ---------- normalize ----------
 @cli.command()
-@click.option("-i", "--input", "i", type=click.Path(exists=True, dir_okay=False, path_type=Path))
+@click.option(
+    "-i", "--input", "i", type=click.Path(exists=True, dir_okay=False, path_type=Path)
+)
 @click.option("-o", "--output", "o", type=click.Path(dir_okay=False, path_type=Path))
 @click.option("-t", "--text", type=str, help="Inline text instead of --input")
 @click.option("--encoding", type=str, help="Force input encoding")
 @click.option("--progress", default=True, is_flag=True, help="Show progress bar")
-def normalize(i: Path | None, o: Path | None, text: str | None, encoding: str | None, progress: bool):
+def normalize(
+    i: Path | None,
+    o: Path | None,
+    text: str | None,
+    encoding: str | None,
+    progress: bool,
+):
     """Normalize text (line-by-line for files)."""
     try:
         from shekar import Normalizer
+
         norm = Normalizer()
     except Exception as e:
-        click.echo(f"Cannot import Normalizer: {e}", err=True); sys.exit(2)
+        click.echo(f"Cannot import Normalizer: {e}", err=True)
+        sys.exit(2)
 
     fout = _open_out(o)
     try:
         lines = list(_iter_lines(i, text, encoding))
-        bar = tqdm(total=len(lines), unit="line", disable=((not progress) or (text is not None)))
+        bar = tqdm(
+            total=len(lines),
+            unit="line",
+            disable=((not progress) or (text is not None)),
+        )
         for s in lines:
             try:
                 out = norm(s)
             except TypeError:
                 out = norm.normalize(s)
-            if fout: fout.write(out + "\n")
-            else: click.echo(out)
+            if fout:
+                fout.write(out + "\n")
+            else:
+                click.echo(out)
             bar.update(1)
         bar.close()
     finally:
-        if fout: fout.close()
+        if fout:
+            fout.close()
+
+
+# ---------- wordcloud ----------
+@cli.command()
+@click.option(
+    "-i",
+    "--input",
+    "i",
+    type=click.Path(exists=True, dir_okay=False, readable=True, path_type=Path),
+    help="Input text file",
+)
+@click.option("-t", "--text", type=str, help="Inline text instead of --input")
+@click.option(
+    "-o",
+    "--output",
[…]
