### 2025-09-21 8dd557a
add tests for cli
