### 2025-09-21 2abde46
Update test_cli.py
