### 2025-09-21 fe0cdec
CLI: Add bidi reshaper option to wordcloud
