### 2025-09-21 a71e239
Add CLI documents + examples
