### 2025-09-21 d698c91
Add CLI docs
