+++ mkdocs.yml
@@ -27,6 +27,7 @@ nav:
     - Overview: index.md
     - Installation: getting_started/installation.md
     - Quick Start: getting_started/quick_start.md
+    - CLI: tutorials/cli.md
     - Preprocessing: tutorials/preprocessing.md
     - Pipeline: tutorials/pipeline.md
     - Tokenization: tutorials/tokenization.md
