### 2025-09-23 3b9682a
Add Persian(❤️) documentations
