+++ docs/fa/index.md
@@ -1,0 +1,603 @@
+
+![Shekar](https://amirivojdan.io/wp-content/uploads/2025/01/shekar-lib.png)
+
+<p align="center">
+<a href="https://pypi.python.org/pypi/shekar" target="_blank"><img alt="PyPI - Version" src="https://img.shields.io/pypi/v/shekar?color=00A693"></a>
+<a href="https://pypi.python.org/pypi/shekar" target="_blank"><img alt="GitHub Actions Workflow Status" src="https://img.shields.io/github/actions/workflow/status/amirivojdan/shekar/test.yml?color=00A693"></a>
+<a href="https://pypi.python.org/pypi/shekar" target="_blank"><img alt="Codecov" src="https://img.shields.io/codecov/c/github/amirivojdan/shekar?color=00A693"></a>
+<a href="https://pypi.python.org/pypi/shekar" target="_blank"><img alt="PyPI - License" src="https://img.shields.io/pypi/l/shekar?color=00A693"></a>
+<a href="https://pypi.python.org/pypi/shekar" target="_blank"><img alt="PyPI - Python Version" src="https://img.shields.io/pypi/pyversions/shekar?color=00A693"></a>
+</p>
+
+
+<p align="center">
+    <em>ساده‌سازی پردازش زبان فارسی برای کاربردهای نوین</em>
+</p>
+
+
+<div dir="rtl">
+<b>شکر</b> یک کتابخانهٔ پایتون برای پردازش زبان فارسی است که نام خود را از داستان طنز <b>«فارسی شکر است»</b> وام گرفته است؛ اثری ماندگار که در سال ۱۹۲۱ به قلم محمدعلی جمالزاده منتشر شد. این داستان به یکی از ارکان نوزایی ادبی ایران بدل شد که با ترویج زبانی ساده و روان، مسیر تازه‌ای در ادبیات معاصر گشود. 
+کتابخانهٔ <b>شکر</b> نیز با الهام از همین نگرش، تلاش می‌کند ابزارهایی کاربردی، ساده و در عین حال دقیق برای پردازش متن فارسی فراهم کند تا پژوهشگران، توسعه‌دهندگان و علاقه‌مندان بتوانند به‌راحتی از آن در پروژه‌های خود استفاده کنند.
+</div>
+
+**فهرست مطالب**
+
+- [راه‌اندازی](#راهاندازی)
+- [پیش‌پردازش](#پیشپردازش)
+  - [یکنواخت‌سازی متن](#یکنواختسازی-متن)
+  - [پردازش دسته‌ای](#پردازش-دستهای)
+  - [پشتیبانی از دکوراتور](#پشتیبانی-از-دکوراتور)
+  - [سفارشی‌سازی](#سفارشیسازی)
+    - [نمای کلی اجزا](#نمای-کلی-اجزا)
+      - [غربال‌ها / پاک‌کننده‌ها](#غربالها--پاککنندهها)
+      - [یکنواخت‌سازی](#یکنواختسازی)
+      - [پوشاننده‌ها](#پوشانندهها)
+  - [بهره‌گیری از زنجیره‌های پردازش](#بهرهگیری-از-زنجیرههای-پردازش)
+- [بخش‌بندی متن](#بخشبندی-متن)
+  - [بخش‌بندی واژگانی](#بخشبندی-واژگانی)
+  - [بخش‌بندی جمله‌ای](#بخشبندی-جملهای)
+- [(Embeddings) بازنمایی](#embeddings-بازنمایی)
+  - [بازنمایی واژگانی](#بازنمایی-واژگانی)
[…]
