### 2025-09-23 5e2d5e6
Add Persian(❤️) documentation
