+++ docs/fa/index.md
@@ -36,11 +36,11 @@
 - [بخش‌بندی متن](#بخشبندی-متن)
   - [بخش‌بندی واژگانی](#بخشبندی-واژگانی)
   - [بخش‌بندی جمله‌ای](#بخشبندی-جملهای)
-- [(Embeddings) بازنمایی](#embeddings-بازنمایی)
+- [(Embeddings) بازنمایی‌ها](#embeddings-بازنماییها)
   - [بازنمایی واژگانی](#بازنمایی-واژگانی)
   - [بازنمایی جمله‌ای](#بازنمایی-جملهای)
 - [ریشه‌یابی سطحی](#ریشهیابی-سطحی)
-- [ریشه‌یابی پایه‌ای](#ریشهیابی-پایهای)
+- [ریشه‌یابی بنیادی](#ریشهیابی-بنیادی)
 - [برچسب‌گذاری نقش‌های دستوری](#برچسبگذاری-نقشهای-دستوری)
 - [شناسایی موجودیت‌های نامدار](#شناسایی-موجودیتهای-نامدار)
 - [کلیدواژه‌یابی](#کلیدواژهیابی)
@@ -224,7 +224,7 @@ for sentence in sentences:
 ما می‌توانیم با هم کار کنیم.
 ```
 
-## (Embeddings) بازنمایی
+## (Embeddings) بازنمایی‌ها
 
 [![Notebook](https://img.shields.io/badge/Notebook-Jupyter-00A693.svg)](examples/embeddings.ipynb)  [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/amirivojdan/shekar/blob/main/examples/embeddings.ipynb)
 
@@ -298,7 +298,7 @@ print(stemmer("خانه‌هایی"))
 خانه
 ```
 
-## ریشه‌یابی پایه‌ای
+## ریشه‌یابی بنیادی
 
 کلاس `Lemmatizer` واژه‌های فارسی را به صورت پایه و فرهنگ‌نامه‌ای آن‌ها نگاشت می‌کند. بر خلاف ریشه‌یابی سطحی که تنها پسوندها را حذف می‌کند، ریشه‌یابی پایه‌ای از قواعد صرف فعل، جست‌وجو در واژگان، و ریشه‌یاب پشتیبان بهره می‌گیرد تا ریشه معتبر تولید کند. این ویژگی باعث می‌شود در کارهایی مانند برچسب‌گذاری نقش‌های دستوری (POS tagging)، نرمال‌سازی متن، و تحلیل‌های زبانی که نیاز به شکل معیار واژه دارند، دقت بیشتری حاصل شود.
 
