+++ README.md
@@ -54,12 +54,25 @@ The story became a cornerstone of Iran's literary renaissance, advocating for ac
 
 ## Installation
 
-To install the package, you can use **`pip`**. Run the following command:
+You can install Shekar with pip. By default, the `CPU` runtime of ONNX is included.
 
 <!-- termynal -->
 ```bash
 $ pip install shekar
 ```
+
+If you want `GPU` acceleration, install with the gpu extra:
+
+<!-- termynal -->
+```bash
+$ pip install "shekar[gpu]"
+```
+
+**Notes:**
+
+- The GPU extra installs onnxruntime-gpu (not available for macOS).
+- If you are unsure which to pick, start with the default CPU install.
+- Both CPU and GPU versions expose the same API, only performance differs depending on your hardware.
 
 ## Preprocessing
 
