+++ docs/fa/index.md
@@ -56,6 +56,8 @@
 
 جهت نصب کتابخانهٔ **شکر**، ابتدا لازم است اطمینان حاصل کنید که `Python` و `pip` بر روی سیستم شما نصب شده باشند. سپس ترمینال یا خط فرمان را باز کرده و دستور زیر را اجرا نمایید. با این کار، بسته از مخزن PyPI دریافت و بر روی سیستم شما نصب خواهد شد. پس از تکمیل فرآیند نصب، می‌توانید از قابلیت‌های کتابخانه در پروژه‌های پردازش زبان فارسی استفاده کنید.
 
+به‌صورت پیش‌فرض، نسخهٔ CPU کتابخانهٔ ONNX نصب می‌شود.
+
 <div dir="ltr">
 <!-- termynal -->
 ```bash
@@ -64,6 +66,23 @@ $ pip install shekar
 Successfully installed shekar!
 ```
 </div>
+
+اگر قصد دارید از شتاب‌دهندهٔ GPU استفاده کنید، بسته را به‌صورت زیر نصب کنید:
+
+<div dir="ltr">
+<!-- termynal -->
+```bash
+$ pip install "shekar[gpu]"
+---> 100%
+Successfully installed shekar!
+```
+</div>
+
+**توجه**
+
+- گزینهٔ `gpu` باعث نصب `onnxruntime-gpu` می‌شود (این نسخه در macOS در دسترس نیست).  
+- اگر مطمئن نیستید کدام گزینه مناسب است، با نصب پیش‌فرض CPU شروع کنید.  
+- هر دو نسخه (CPU و GPU) رابط برنامه‌نویسی یکسانی دارند؛ تنها تفاوت در کارایی بسته به سخت‌افزار شماست.
 
 ## پیش‌پردازش
 
