### 2025-09-29 f1cf29a
Exclude GPU dependencies on test
