+++ .github/workflows/test.yml
@@ -31,8 +31,8 @@ jobs:
       - name: Upgrade hatch
         run: uv pip install --upgrade hatch hatchling
 
-      - name: Install Dependencies
-        run: uv pip install -r pyproject.toml --all-extras
+      - name: Install base + dev dependencies
+        run: uv pip install -e ".[dev]"
 
       - name: Install Ruff
         run: uv pip install ruff
