### 2025-09-29 f1a40e2
Add GPU support
