+++ shekar/utils.py
@@ -1,4 +1,5 @@
 import re
+import onnxruntime as ort
 
 
 def is_informal(text, threshold=1) -> bool:
@@ -36,3 +37,26 @@ def is_informal(text, threshold=1) -> bool:
 
     classification = True if match_count >= threshold else False
     return classification
+
+
+def get_onnx_providers() -> list[str]:
+    """
+    Get the list of available ONNX Runtime execution providers, prioritizing GPU providers if available.
+    This function checks for the presence of various execution providers and returns a list ordered by preference.
+    Returns:
+        list: A list of available ONNX Runtime execution providers ordered by preference.
+    """
+
+    PREFERRED = [
+        "TensorrtExecutionProvider",  # NVIDIA TensorRT
+        "CUDAExecutionProvider",  # NVIDIA CUDA
+        "ROCMExecutionProvider",  # AMD ROCm (Linux)
+        "DmlExecutionProvider",  # Windows DirectML
+        "OpenVINOExecutionProvider",  # Intel CPU/iGPU
+        "CoreMLExecutionProvider",  # macOS
+        "CPUExecutionProvider",  # always last
+    ]
+
+    available = ort.get_available_providers()
+    providers = [ep for ep in PREFERRED if ep in available]
+    return providers
