+++ shekar/embeddings/albert_embedder.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 import onnxruntime
 import numpy as np
-
 from shekar.hub import Hub
 from .base import BaseEmbedder
 from shekar.tokenization import AlbertTokenizer
+from shekar.utils import get_onnx_providers
 
 
 class AlbertEmbedder(BaseEmbedder):
@@ -13,8 +13,9 @@ class AlbertEmbedder(BaseEmbedder):
         resource_name = "albert_persian_mlm_embeddings.onnx"
         if model_path is None or not Path(model_path).exists():
             model_path = Hub.get_resource(file_name=resource_name)
-
-        self.session = onnxruntime.InferenceSession(model_path)
+        self.session = onnxruntime.InferenceSession(
+            model_path, providers=get_onnx_providers()
+        )
         self.tokenizer = AlbertTokenizer(enable_padding=True, enable_truncation=True)
         self.vector_size = 768
 
