### 2025-10-01 #8860 @crvernon
That is correct @amirivojdan - knowledge of the Persian language would not be necessary to conduct this review.  @evamaxfield your expertise is valued in your evaluation of the software via a checklist that you can generate when I start the full review shortly.  Thank you!
