### 2025-10-01 #9128 @editorialbot
[REVIEW]: Shekar: A Python Toolkit for Persian Natural Language Processing
**Submitting author:** @amirivojdan (<a href="http://orcid.org/0000-0003-3741-3979">Ahmad Amirivojdan</a>)
**Repository:** https://github.com/amirivojdan/shekar
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
