### 2025-10-01 #9128 @linuxscout
@amirivojdan:
The README file is too long. It would be more effective to keep it concise by summarizing the project’s features, installation steps, and a few examples, while moving detailed features and extended examples to a separate file or to a dedicated ReadTheDocs documentation.
