### 2025-10-01 #9128 @linuxscout
@amirivojdan:
Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
