+++ README.md
@@ -14,14 +14,9 @@
     <em>Simplifying Persian NLP for Modern Applications</em>
 </p>
 
-**Shekar** (meaning 'sugar' in Persian) is a Python library for Persian natural language processing, named after the influential satirical story *"فارسی شکر است"* (Persian is Sugar) published in 1921 by Mohammad Ali Jamalzadeh.
-The story became a cornerstone of Iran's literary renaissance, advocating for accessible yet eloquent expression. Shekar embodies this philosophy in its design and development.
+**Shekar** is an open-source Python library for Persian natural language processing (NLP). It provides tools for text preprocessing, tokenization, part-of-speech tagging, named entity recognition, embeddings, spell checking, and more. With its modular pipeline design, Shekar makes it easy to build reproducible workflows for both research and production applications.
 
-<div dir="rtl">
-«شِکَر» یک کتابخانهٔ پایتون برای پردازش زبان فارسی است که نام خود را از داستان طنز «فارسی شکر است» وام گرفته است؛ اثری ماندگار که در سال ۱۹۲۱ به قلم محمدعلی جمالزاده منتشر شد. این داستان به یکی از ارکان نوزایی ادبی ایران بدل شد که با ترویج زبانی ساده و روان، مسیر تازه‌ای در ادبیات معاصر گشود. 
-کتابخانهٔ «شِکَر» نیز با الهام از همین نگرش، تلاش می‌کند ابزارهایی کاربردی، ساده و در عین حال دقیق برای پردازش متن فارسی فراهم کند تا پژوهشگران، توسعه‌دهندگان و علاقه‌مندان بتوانند به‌راحتی از آن در پروژه‌های خود استفاده کنند.
-</div>
-
+📖 Documentation: https://lib.shekar.io/
 
 ### Table of Contents
 
@@ -30,11 +25,7 @@ The story became a cornerstone of Iran's literary renaissance, advocating for ac
   - [GPU Acceleration (NVIDIA CUDA)](#gpu-acceleration-nvidia-cuda)
 - [Preprocessing](#preprocessing)
   - [Normalizer](#normalizer)
-  - [Batch Processing](#batch-processing)
-  - [Decorator Support](#decorator-support)
   - [Customization](#customization)
-    - [Component Overview](#component-overview)
-  - [Using Pipelines](#using-pipelines)
 - [Tokenization](#tokenization)
   - [WordTokenizer](#wordtokenizer)
   - [SentenceTokenizer](#sentencetokenizer)
@@ -45,12 +36,10 @@ The story became a cornerstone of Iran's literary renaissance, advocating for ac
 - [Lemmatization](#lemmatization)
 - [Part-of-Speech Tagging](#part-of-speech-tagging)
 - [Named Entity Recognition (NER)](#named-entity-recognition-ner)
-  - [Chaining with Pipelines](#chaining-with-pipelines)
 - [Keyword Extraction](#keyword-extraction)
 - [Spell Checking](#spell-checking)
 - [WordCloud](#wordcloud)
 - [Command-Line Interface (CLI)](#command-line-interface-cli)
-  - [Commands](#commands)
 - [Download Models](#download-models)
 
 
@@ -102,87 +91,9 @@ print(normalizer(text))
 «فارسی شکر است» نام داستان کوتاه طنزآمیزی از محمد‌علی جمالزاده‌ی گرامی می‌باشد که در سال ۱۹۲۱ منتشر شده‌است و آغازگر تحول بزرگی در ادبیات معاصر ایران به شمار می‌رود.
 ```
 
-### Batch Processing
-
[…]
