### 2025-10-06 b6fb49a
Refactor filters to maskers and update API
Moved and renamed all filter classes to maskers, updating their names and constructor arguments for consistency (e.g., 'replace_with' to 'mask_token'). Updated imports, aliases, and __all__ lists in __init__.py files to reflect these changes. Also moved repeated letter filter to normalizers as RepeatedLetterNormalizer. Incremented version to 0.1.30.
