+++ shekar/preprocessing/__init__.py
@@ -1,14 +1,16 @@
-from .filters import (
-    DiacriticFilter,
-    DigitFilter,
-    EmojiFilter,
-    HashtagFilter,
-    HTMLTagFilter,
-    MentionFilter,
-    NonPersianLetterFilter,
-    PunctuationFilter,
-    RepeatedLetterFilter,
-    StopWordFilter,
+from .maskers import (
+    DiacriticMasker,
+    DigitMasker,
+    EmojiMasker,
+    HashtagMasker,
+    HTMLTagMasker,
+    MentionMasker,
+    NonPersianLetterMasker,
+    PunctuationMasker,
+    StopWordMasker,
+    EmailMasker,
+    URLMasker,
+    OffensiveWordMasker,
     # aliases
     DiacriticRemover,
     EmojiRemover,
@@ -18,8 +20,10 @@ from .filters import (
     HashtagRemover,
     MentionRemover,
     DigitRemover,
-    RepeatedLetterRemover,
-    HTMLRemover,
+    HTMLTagRemover,
+    EmailRemover,
+    URLRemover,
+    OffensiveWordRemover,
     # action-based aliases
     RemoveDiacritics,
     RemoveEmojis,
@@ -29,15 +33,23 @@ from .filters import (
     RemoveHashtags,
     RemoveMentions,
     RemoveDigits,
-    RemoveRepeatedLetters,
     RemoveHTMLTags,
-)
-
-from .maskers import (
-    EmailMasker,
-    URLMasker,
+    RemoveEmails,
+    RemoveURLs,
+    RemoveOffensiveWords,
+    # Maskers
     MaskEmails,
     MaskURLs,
+    MaskEmojis,
+    MaskDigits,
+    MaskPunctuations,
+    MaskNonPersianLetters,
+    MaskStopWords,
+    MaskHashtags,
+    MaskMentions,
+    MaskDiacritics,
+    MaskHTMLTags,
+    MaskOffensiveWords,
 )
 
 from .normalizers import (
@@ -53,20 +65,25 @@ from .normalizers import (
     NormalizeSpacings,
     YaNormalizer,
     NormalizeYas,
+    RepeatedLetterNormalizer,
+    NormalizeRepeatedLetters,
 )
 
 __all__ = [
-    # Filters
-    "DiacriticFilter",
-    "DigitFilter",
-    "EmojiFilter",
-    "HashtagFilter",
-    "HTMLTagFilter",
-    "MentionFilter",
-    "NonPersianLetterFilter",
-    "PunctuationFilter",
-    "RepeatedLetterFilter",
-    "StopWordFilter",
+    # Maskers
+    "DiacriticMasker",
+    "EmojiMasker",
+    "NonPersianLetterMasker",
+    "PunctuationMasker",
+    "StopWordMasker",
+    "HashtagMasker",
+    "MentionMasker",
+    "DigitMasker",
+    "RepeatedLetterMasker",
+    "HTMLTagMasker",
+    "EmailMasker",
+    "URLMasker",
+    "OffensiveWordMasker",
     # aliases
     "DiacriticRemover",
     "EmojiRemover",
@@ -76,8 +93,10 @@ __all__ = [
     "HashtagRemover",
     "MentionRemover",
     "DigitRemover",
-    "RepeatedLetterRemover",
-    "HTMLRemover",
+    "HTMLTagRemover",
+    "EmailRemover",
+    "URLRemover",
+    "OffensiveWordRemover",
     # action-based aliases
     "RemoveDiacritics",
     "RemoveEmojis",
@@ -87,24 +106,36 @@ __all__ = [
     "RemoveHashtags",
     "RemoveMentions",
[…]
