+++ shekar/normalizer.py
@@ -5,13 +5,13 @@ from shekar.preprocessing import (
     AlphabetNormalizer,
     DigitNormalizer,
     SpacingNormalizer,
-    EmojiFilter,
-    EmailMasker,
-    URLMasker,
-    DiacriticFilter,
-    # NonPersianLetterFilter,
-    HTMLTagFilter,
-    RepeatedLetterFilter,
+    MaskEmojis,
+    MaskEmails,
+    MaskURLs,
+    RemoveDiacritics,
+    # NonPersianLetterMasker,
+    MaskHTMLTags,
+    RepeatedLetterNormalizer,
     ArabicUnicodeNormalizer,
     YaNormalizer,
 )
@@ -25,15 +25,15 @@ class Normalizer(Pipeline):
                 ("ArabicUnicodeNormalizer", ArabicUnicodeNormalizer()),
                 ("DigitNormalizer", DigitNormalizer()),
                 ("PunctuationNormalizer", PunctuationNormalizer()),
-                ("EmailMasker", EmailMasker(mask="")),
-                ("URLMasker", URLMasker(mask="")),
-                ("EmojiFilter", EmojiFilter()),
-                ("HTMLTagFilter", HTMLTagFilter()),
-                ("DiacriticFilter", DiacriticFilter()),
-                ("RepeatedLetterFilter", RepeatedLetterFilter()),
+                ("EmailMasker", MaskEmails(mask_token=" ")),
+                ("URLMasker", MaskURLs(mask_token=" ")),
+                ("EmojiMasker", MaskEmojis(mask_token=" ")),
+                ("HTMLTagMasker", MaskHTMLTags(mask_token=" ")),
+                ("DiacriticRemover", RemoveDiacritics()),
+                ("RepeatedLetterNormalizer", RepeatedLetterNormalizer()),
                 # ("NonPersianLetterFilter", NonPersianLetterFilter()),
                 ("SpacingNormalizer", SpacingNormalizer()),
-                ("YaNormalizer", YaNormalizer()),
+                ("YaNormalizer", YaNormalizer(style="joda")),
             ]
         super().__init__(steps=steps)
 
