+++ tests/unit/preprocessing/test_preprocessors.py
@@ -6,19 +6,20 @@ from shekar.preprocessing import (
     DigitNormalizer,
     SpacingNormalizer,
     YaNormalizer,
-    EmojiFilter,
+    EmojiMasker,
     EmailMasker,
     URLMasker,
-    DiacriticFilter,
-    NonPersianLetterFilter,
-    HTMLTagFilter,
-    RepeatedLetterFilter,
+    DiacriticRemover,
+    NonPersianLetterMasker,
+    HTMLTagMasker,
+    RepeatedLetterNormalizer,
     ArabicUnicodeNormalizer,
-    StopWordFilter,
-    PunctuationFilter,
-    DigitFilter,
-    MentionFilter,
-    HashtagFilter,
+    StopWordRemover,
+    PunctuationRemover,
+    DigitRemover,
+    MentionMasker,
+    HashtagMasker,
+    OffensiveWordMasker,
 )
 
 from shekar.transforms import (
@@ -128,7 +129,7 @@ def test_ya_normalizer():
 
 
 def test_mask_email():
-    email_masker = EmailMasker(mask="")
+    email_masker = EmailMasker(mask_token="")
 
     input_text = "ایمیل من: she.kar@shekar.panir.io"
     expected_output = "ایمیل من:"
@@ -140,7 +141,7 @@ def test_mask_email():
 
 
 def test_mask_url():
-    url_masker = URLMasker(mask="")
+    url_masker = URLMasker(mask_token="")
 
     input_text = "لینک: https://shekar.parsi-shekar.com"
     expected_output = "لینک:"
@@ -247,7 +248,7 @@ def test_unify_arabic_unicode():
 
 
 def test_remove_punctuations():
-    punc_Filter = PunctuationFilter()
+    punc_Filter = PunctuationRemover()
 
     input_text = "اصفهان، نصف جهان!"
     expected_output = "اصفهان نصف جهان"
@@ -259,7 +260,7 @@ def test_remove_punctuations():
 
 
 def test_remove_redundant_characters():
-    redundant_character_Filter = RepeatedLetterFilter()
+    redundant_character_Filter = RepeatedLetterNormalizer()
     input_text = "سلامم"
     expected_output = "سلامم"
     assert redundant_character_Filter(input_text) == expected_output
@@ -274,7 +275,7 @@ def test_remove_redundant_characters():
 
 
 def test_remove_emojis():
-    emoji_Filter = EmojiFilter()
+    emoji_Filter = EmojiMasker()
     input_text = "😊🇮🇷سلام گلای تو خونه!🎉🎉🎊🎈"
     expected_output = "سلام گلای تو خونه!"
     assert emoji_Filter(input_text) == expected_output
@@ -286,7 +287,7 @@ def test_remove_emojis():
 
 
 def test_remove_diacritics():
-    diacritics_Filter = DiacriticFilter()
+    diacritics_Filter = DiacriticRemover()
     input_text = "مَنْ"
     expected_output = "من"
     assert diacritics_Filter(input_text) == expected_output
@@ -297,7 +298,7 @@ def test_remove_diacritics():
 
 
 def test_remove_stopwords():
-    stopword_Filter = StopWordFilter()
+    stopword_Filter = StopWordRemover()
     input_text = "این یک جملهٔ نمونه است"
     expected_output = "جملهٔ نمونه"
     assert stopword_Filter(input_text) == expected_output
@@ -310,14 +311,14 @@ def test_remove_stopwords():
     expected_output = "ایران"
     assert stopword_Filter(input_text) == expected_output
 
-    stopword_Filter = StopWordFilter(replace_with="|")
+    stopword_Filter = StopWordRemover(mask_token="|")
[…]
