### 2025-10-06 d597715
Docs: update component lists and improve structure
Updated English and Persian documentation to clarify and reorganize the lists of normalizers and maskers, removing references to filters/removers and adding new components. Improved table formatting, added details for each component, and enhanced section headings for better readability and consistency across languages.
