+++ docs/fa/index.md
@@ -31,9 +31,8 @@
   - [پشتیبانی از دکوراتور](#پشتیبانی-از-دکوراتور)
   - [سفارشی‌سازی](#سفارشیسازی)
     - [نمای کلی اجزا](#نمای-کلی-اجزا)
-      - [غربال‌ها / پاک‌کننده‌ها](#غربالها--پاککنندهها)
-      - [یکنواخت‌سازی](#یکنواختسازی)
-      - [پوشاننده‌ها](#پوشانندهها)
+      - [یکنواخت‌سازها (Normalizers)](#یکنواختسازها-normalizers)
+      - [پوشاننده‌ها (Maskers)](#پوشانندهها-maskers)
   - [بهره‌گیری از زنجیره‌های پردازش](#بهرهگیری-از-زنجیرههای-پردازش)
 - [بخش‌بندی متن](#بخشبندی-متن)
   - [بخش‌بندی واژگانی](#بخشبندی-واژگانی)
@@ -153,43 +152,38 @@ print(process_text("تو را من چشم👀 در راهم!"))
 
 ### سفارشی‌سازی
 
-برای سفارشی‌سازی پیشرفته، کتابخانهٔ شِکَر یک چارچوب ماژولار و ترکیب‌پذیر برای پیش‌پردازش متن ارائه می‌دهد. این چارچوب شامل مؤلفه‌هایی مانند filters، normalizers و maskers است که می‌توانند به‌صورت مستقل به‌کار گرفته شوند یا با استفاده از کلاس Pipeline و عملگر `|` به‌طور انعطاف‌پذیر با یکدیگر ترکیب شوند.
+برای سفارشی‌سازی پیشرفته، کتابخانهٔ شِکَر یک چارچوب ماژولار و ترکیب‌پذیر برای پیش‌پردازش متن ارائه می‌دهد. این چارچوب شامل مؤلفه‌هایی مانند `normalizers` و `maskers` است که می‌توانند به‌صورت مستقل به‌کار گرفته شوند یا با استفاده از کلاس `Pipeline` و عملگر `|` به‌طور انعطاف‌پذیر با یکدیگر ترکیب شوند.
 
 #### نمای کلی اجزا
 
-##### غربال‌ها / پاک‌کننده‌ها
+##### یکنواخت‌سازها (Normalizers)
 
 | مؤلفه | نام‌های دیگر | شرح |
 |----------|---------|-------------|
-| `DiacriticFilter` | `DiacriticRemover`, `RemoveDiacritics` | نشانه‌های حرکتی (اعراب) در متون فارسی/عربی را پاک می‌کند. |
-| `EmojiFilter` | `EmojiRemover`, `RemoveEmojis` | نشانه‌ها/شکلک‌های ایموجی را از متن پاک می‌کند. |
-| `NonPersianLetterFilter` | `NonPersianRemover`, `RemoveNonPersianLetters` | همهٔ داده‌های غیر‌فارسی را حذف کرده و در صورت تنظیم می‌تواند محتوای انگلیسی را حفظ کند. |
-| `PunctuationFilter` | `PunctuationRemover`, `RemovePunctuations` | کلیهٔ نشانه‌های سجاوندی را از متن پاک می‌کند. |
-| `StopWordFilter` | `StopWordRemover`, `RemoveStopWords` |  ایست‌واژه‌های پرتکرار در زبان فارسی را از متن پاک می‌کند. |
-| `DigitFilter` | `DigitRemover`, `RemoveDigits` | کلیهٔ نویسه‌های عددی را از متن پاک می‌کند. |
[…]
