### 2025-10-08 72ad16b
Add sentiment analysis module
Introduces a sentiment analysis module for Persian text using a quantized ALBERT model fine-tuned for binary sentiment classification. Updates documentation and README with usage examples, adds model hash to hub, and exposes SentimentClassifier in the package API.
