+++ shekar/hub.py
@@ -5,6 +5,7 @@ import hashlib
 
 MODEL_HASHES = {
     "albert_persian_tokenizer.json": "79716aa7d8aeee80d362835da4f33e2b36b69fe65c257ead32c5ecd850e9ed17",
+    "albert_persian_sentiment_binary_q8.onnx": "377c322edc3c0de0c48bf3fd4420c7385158bd34492f5b157ea6978745c50e4a",
     "albert_persian_ner_q8.onnx": "a3d2b1d2c167abd01e6b663279d3f8c3bb1b3d0411f693515cd0b31a5a3d3e80",
     "albert_persian_pos_q8.onnx": "8b5a2761aae83911272763034e180345fe12b2cd45b6de0151db9fbf9d3d8b31",
     "albert_persian_mlm_embeddings.onnx": "6b2d987ba409fd6957764742e30bfbbe385ab33c210caeb313aa9a2eb9afa51a",
@@ -78,3 +79,13 @@ class Hub:
         except Exception as e:
             print(f"Error downloading the file: {e}")
             return False
+
+
+if __name__ == "__main__":
+    import sys
+
+    if len(sys.argv) != 2:
+        print("Usage: python hub.py <file_path>")
+        sys.exit(1)
+    file_path = sys.argv[1]
+    print(Hub.compute_sha256_hash(file_path))
