### 2025-10-08 92f10a5
Add SentimentClassifier to public API
Included SentimentClassifier in the __all__ list to make it publicly accessible. Also made minor docstring formatting improvements in the SentimentClassifier class.
