+++ shekar/sentiment_analysis/base.py
@@ -7,12 +7,13 @@ SENTIMENT_REGISTRY = {
 
 
 class SentimentClassifier(BaseTransform):
-    """ A wrapper class for sentiment analysis models. 
+    """A wrapper class for sentiment analysis models.
     Currently, it supports only the "albert-binary" model.
      Args:
         model (str): The sentiment analysis model to use. Default is "albert-binary".
         model_path (str, optional): Path to a custom model file. If None, the default model will be used.
     """
+
     def __init__(self, model: str = "albert-binary", model_path=None):
         model = model.lower()
         if model not in SENTIMENT_REGISTRY:
@@ -23,7 +24,7 @@ class SentimentClassifier(BaseTransform):
         self.model = SENTIMENT_REGISTRY[model](model_path=model_path)
 
     def transform(self, X: str) -> list:
-        """ Perform sentiment analysis on the input text.
+        """Perform sentiment analysis on the input text.
         Args:
             X (str): Input text.
             Returns:
