### 2025-10-08 a347a21
Add Sentiment Analysis to Persian(❤️) docs
Introduced a new section on sentiment analysis in the Persian documentation, including usage instructions and code examples for the SentimentClassifier module.
