+++ docs/fa/index.md
@@ -44,6 +44,7 @@
 - [ریشه‌یابی بنیادی](#ریشهیابی-بنیادی)
 - [برچسب‌گذاری نقش‌های دستوری](#برچسبگذاری-نقشهای-دستوری)
 - [شناسایی موجودیت‌های نامدار](#شناسایی-موجودیتهای-نامدار)
+- [سنجش احساسات](#سنجش-احساسات)
 - [کلیدواژه‌یابی](#کلیدواژهیابی)
 - [غلط‌یابی املایی](#غلطیابی-املایی)
 - [ابر واژگان](#ابر-واژگان)
@@ -445,6 +446,27 @@ for text, label in entities:
 ```
 
 این زنجیره‌سازی کدی تمیز و خوانا فراهم می‌کند و به شما امکان می‌دهد زنجیره‌های NLP سفارشی بسازید که در یک مرحله شامل پیش‌پردازش و برچسب‌گذاری باشند.
+
+
+## سنجش احساسات
+
+ماژول `SentimentClassifier` امکان واکاوی خودکار احساسات در متون فارسی را با استفاده از مدل‌های مبتنی بر مدل‌‌های زبانی فراهم می‌کند. در حال حاضر از مدل `AlbertBinarySentimentClassifier` پشتیبانی می‌کند، مدلی سبک بر پایهٔ ALBERT که بر روی دادگان Snapfood آموزش دیده و نوشته را به دو دستهٔ مثبت یا منفی طبقه‌بندی می‌کند و علاوه بر برچسب پیش‌بینی‌شده، امتیاز اطمینان آن را نیز برمی‌گرداند.
+
+**نمونه کد:**
+
+```python
+from shekar import SentimentClassifier
+
+sentiment_classifier = SentimentClassifier()
+
+print(sentiment_classifier("سریال قصه‌های مجید عالی بود!"))
+print(sentiment_classifier("فیلم ۳۰۰ افتضاح بود!"))
+```
+
+```output
+('positive', 0.9923112988471985)
+('negative', 0.9330866932868958)
+```
 
 ## کلیدواژه‌یابی
 
