### 2025-10-14 4cd6e42
Add offensive language detection
Introduces a new toxicity module with an offensive language classifier using ONNX.
