### 2025-10-14 58de239
Add minor formatting improvements to toxicity module
