### 2025-10-14 cd3bf51
Refactor offensive classifier registry and improve docs
