+++ shekar/toxicity/logistic_offensive_classifier.py
@@ -8,6 +8,19 @@ from shekar.preprocessing import StopWordRemover
 
 
 class LogisticOffensiveClassifier(BaseTransform):
+    """Logistic model for offensive language detection.
+    This model is trained on Naseza(ناسزا) Persian offensive language dataset.
+    Args:
+        model_path (str | Path, optional): Path to a custom model file. If None, the default model will be used.
+
+    Example:
+        >>> model = LogisticOffensiveClassifier()
+        >>> model.transform("این یک متن معمولی است.")
+        ('neutral', 0.987654321)
+        >>> model.transform("تو خیلی احمق و بی‌شرفی!")
+        ('offensive', 0.9987654321)
+    """
+
     def __init__(self, model_path: str | Path = None):
         super().__init__()
         resource_name = "tfidf_logistic_offensive.onnx"
@@ -21,7 +34,7 @@ class LogisticOffensiveClassifier(BaseTransform):
         self.id2label = {0: "neutral", 1: "offensive"}
         self.stopword_remover = StopWordRemover()
 
-    def transform(self, X: str) -> str:
+    def transform(self, X: str) -> tuple:
         X = self.stopword_remover(X)
 
         in_name = self.session.get_inputs()[0].name
