### 2025-10-14 f8799a3
Add toxicity detection section to README
Added documentation for the new toxicity detection feature, including usage examples and model download links. Updated the table of contents and model download section to include the Offensive Language Classifier.
