### 2025-10-14 #9128 @evamaxfield
Hello everyone! I have no reviewed the library and the paper! Everything looks really great, nice work @amirivojdan. I have one very minor PR from something I noticed in the README while testing: https://github.com/amirivojdan/shekar/pull/9

Otherwise, everything else is checked off.
