### 2025-10-15 bb8dd9f
Add references and details for sentiment and toxicity modules
