+++ paper.bib
@@ -71,3 +71,26 @@
   publisher={Wiley Online Library},
   doi={10.1002/9780470689646.ch1}
 }
+
+@article{farahani2021parsbert,
+  title={Parsbert: Transformer-based model for persian language understanding},
+  author={Farahani, Mehrdad and Gharachorloo, Mohammad and Farahani, Marzieh and Manthouri, Mohammad},
+  journal={Neural Processing Letters},
+  volume={53},
+  number={6},
+  pages={3831--3847},
+  year={2021},
+  publisher={Springer},
+  doi={10.1007/s11063-021-10528-4}
+}
+
+@dataset{amirivojdan_2025_naseza,
+  author       = {Ahmad Amirivojdan},
+  title        = {Naseza: A Large-Scale Dataset for Persian Hate Speech and Offensive Language Detection},
+  year         = {2025},
+  publisher    = {Zenodo},
+  version      = {v1.0.0},
+  doi          = {10.5281/zenodo.17355123},
+  url          = {https://doi.org/10.5281/zenodo.17355123},
+  license      = {CC0-1.0}
+}
