### 2025-10-15 9b5ae68
Add word cloud example to paper
