### 2025-10-21 #9128 @crvernon
👋  @amirivojdan - we are almost there! I just need you to make the following changes in your paper:

- [ ] LINE 49:  Missing space in between "word-level," and "sentence-level"
- [ ] LINE 107:  The "p" in "persian" should be capitalized.  You can do this in your bib file by placing curly brackets around the characters you wish to maintain the formatting of.
- [ ] LINE 113:  The "p" in "persian" should be capitalized.
- [ ] LINE 116:  The "p" in "persian" should be capitalized.
- [ ] LINES 121-122:  "Albert" and "bert" should be capitalized.
- [ ] LINE 125:  The "p" in "persian" should be capitalized.
- [ ] LINE 127:  The "p" in "persian" should be capitalized.
- [ ] LINE 134:  The "f" in "farsi" should be capitalized.

- [ ]  Conduct a GitHub release of the current reviewed version of the software and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) if you haven't already.  Please ensure that the software archive uses the same license as the license you have posted on code repository.

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should match the people and the order in which they are listed on your paper exactly).

- [ ]  Please respond with the DOI of the archived version here along with the version number released.
