### 2025-10-21 7578754
Fix formatting in bibliography and documentation
