+++ paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{jafari2025dadmatools,
-  title={DadmaTools V2: an Adapter-Based Natural Language Processing Toolkit for the Persian Language},
+  title={DadmaTools V2: an Adapter-Based Natural Language Processing Toolkit for the {P}ersian Language},
   author={Jafari, Sadegh and Farsi, Farhan and Ebrahimi, Navid and Sajadi, Mohamad Bagher and Eetemadi, Sauleh},
   booktitle={Proceedings of the 1st Workshop on NLP for Languages Using Arabic Script},
   pages={37--43},
@@ -7,14 +7,14 @@
 }
 
 @inproceedings{mohtaj2018parsivar,
-  title={Parsivar: A language processing toolkit for Persian},
+  title={Parsivar: A language processing toolkit for {P}ersian},
   author={Mohtaj, Salar and Roshanfekr, Behnam and Zafarian, Atefeh and Asghari, Habibollah},
   booktitle={Proceedings of the eleventh international conference on language resources and evaluation (lrec 2018)},
   year={2018}
 }
 
 @article{sabouri2022naab,
-  title={naab: A ready-to-use plug-and-play corpus for Farsi},
+  title={naab: A ready-to-use plug-and-play corpus for {F}arsi},
   author={Sabouri, Sadra and Rahmati, Elnaz and Gooran, Soroush and Sameti, Hossein},
   journal={arXiv preprint arXiv:2208.13486},
   year={2022},
@@ -47,7 +47,7 @@
 }
 
 @article{rasooli2020persian,
-  title={The Persian dependency treebank made universal},
+  title={The {P}ersian dependency treebank made universal},
   author={Rasooli, Mohammad Sadegh and Safari, Pegah and Moloodi, Amirsaeid and Nourian, Alireza},
   journal={arXiv preprint arXiv:2009.10205},
   year={2020},
@@ -55,7 +55,7 @@
 }
 
 @article{lan2019albert,
-  title={Albert: A lite bert for self-supervised learning of language representations},
+  title={{ALBERT}: A lite {BERT} for self-supervised learning of language representations},
   author={Lan, Zhenzhong and Chen, Mingda and Goodman, Sebastian and Gimpel, Kevin and Sharma, Piyush and Soricut, Radu},
   journal={arXiv preprint arXiv:1909.11942},
   year={2019},
@@ -73,7 +73,7 @@
 }
 
 @article{farahani2021parsbert,
-  title={Parsbert: Transformer-based model for persian language understanding},
+  title={Parsbert: Transformer-based model for {P}ersian language understanding},
   author={Farahani, Mehrdad and Gharachorloo, Mohammad and Farahani, Marzieh and Manthouri, Mohammad},
   journal={Neural Processing Letters},
   volume={53},
@@ -86,7 +86,7 @@
 
 @dataset{amirivojdan_2025_naseza,
   author       = {Ahmad Amirivojdan},
-  title        = {Naseza: A Large-Scale Dataset for Persian Hate Speech and Offensive Language Detection},
+  title        = {Naseza: A Large-Scale Dataset for {P}ersian Hate Speech and Offensive Language Detection},
   year         = {2025},
   publisher    = {Zenodo},
   version      = {v1.0.0},
