### 2025-10-21 b8207ee
Release version 1.0.0 and update metadata
