### 2025-10-21 53fc6f4
Add citation section to README
